Basic and OAuth2 security configs
