+++ pom.xml
@@ -49,11 +49,6 @@
         </dependency>
 
         <!--        <dependency>-->
-        <!--            <groupId>org.springframework.security.oauth</groupId>-->
-        <!--            <artifactId>spring-security-oauth2</artifactId>-->
-        <!--        </dependency>-->
-
-        <!--        <dependency>-->
         <!--            <groupId>org.springframework.security</groupId>-->
         <!--            <artifactId>spring-security-jwt</artifactId>-->
         <!--        </dependency>-->
