+++ src/main/java/cane/brothers/sec/BasicUserController.java
@@ -1,10 +1,9 @@
 package cane.brothers.sec;
 
-
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Profile;
+import org.springframework.http.MediaType;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -12,16 +11,16 @@ import org.springframework.web.bind.annotation.RestController;
 import java.security.Principal;
 
 /**
- * @author mniedre
+ * Created by cane
  */
 @Slf4j
-@Profile("security")
+@Profile("basic")
 @RestController
-public class UserController {
+public class BasicUserController {
 
     @GetMapping
-    @RequestMapping("/user")
-    public Principal user(@AuthenticationPrincipal OAuth2User principal) {
-        return () -> principal.getAttribute("login");
+    @RequestMapping(path = "/user", produces = MediaType.APPLICATION_JSON_VALUE)
+    public Principal user(@AuthenticationPrincipal Principal principal) {
+        return principal;
     }
 }
