+++ src/main/java/cane/brothers/sec/OAuth2SecurityConfig.java
@@ -1,19 +1,26 @@
 package cane.brothers.sec;
 
+import org.springframework.boot.autoconfigure.security.oauth2.client.EnableOAuth2Sso;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.web.authentication.HttpStatusEntryPoint;
+import org.springframework.web.cors.CorsConfiguration;
+import org.springframework.web.cors.CorsConfigurationSource;
 import org.springframework.web.cors.CorsUtils;
+import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
+
+import java.util.Arrays;
 
 /**
  * @author mniedre
  */
-@Profile("security")
+@Profile("oauth2")
 @Configuration
-public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
+public class OAuth2SecurityConfig extends WebSecurityConfigurerAdapter {
 
     @Override
     protected void configure(HttpSecurity http) throws Exception {
@@ -31,26 +38,26 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
                 .exceptionHandling(e -> e
                         .authenticationEntryPoint(new HttpStatusEntryPoint(HttpStatus.UNAUTHORIZED)))
                 .oauth2Login(o -> o
-                        .defaultSuccessUrl("/", true)
+                        .defaultSuccessUrl("http://localhost:4200/", true)
                         .failureUrl("/error"))
                 .logout(l -> l
                                 //.logoutRequestMatcher(new AntPathRequestMatcher("/logout"))
-                                .logoutSuccessUrl("/")
-//                        .invalidateHttpSession(true)        // set invalidation state when logout
-                                .deleteCookies("JSESSIONID")
-//                        .clearAuthentication(true)
-                );
-
-//                .cors();
+                                .logoutSuccessUrl("http://localhost:4200/")
+                        .invalidateHttpSession(true)        // set invalidation state when logout
+                        .deleteCookies("JSESSIONID")
+                        .clearAuthentication(true)
+                )
+                .cors();
     }
 
-//    @Bean
-//    CorsConfigurationSource corsConfigurationSource() {
-//        CorsConfiguration configuration = new CorsConfiguration();
-//        configuration.setAllowedOrigins(Arrays.asList("http://localhost:3000", "https://github.com/login/oauth/authorize"));
-//        configuration.setAllowedMethods(Arrays.asList("GET"));
-//        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-//        source.registerCorsConfiguration("/**", configuration);
-//        return source;
-//    }
+    @Bean
+    CorsConfigurationSource corsConfigurationSource() {
+        CorsConfiguration configuration = new CorsConfiguration();
+        configuration.setAllowedOrigins(Arrays.asList("http://localhost:4200"));
+        configuration.setAllowedMethods(Arrays.asList("GET", "POST", "OPTIONS", "DELETE", "PUT"));
+        configuration.setAllowedHeaders(Arrays.asList("authorization"));
+        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
+        source.registerCorsConfiguration("/**", configuration);
+        return source;
+    }
 }
