+++ src/main/resources/application.yml
@@ -1,3 +1,3 @@
 spring:
   profiles:
-    active: dev,db,management #,security,swagger,management
+    active: dev,db,management,oauth2 #,swagger,management
